refactor: TableGroup dto 수정
