+++ src/main/java/kitchenpos/dto/request/order/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.dto.request.order;
 
 import java.util.List;
