+++ src/main/java/kitchenpos/dto/request/product/CreateProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.dto.request.product;
 
 import java.math.BigDecimal;
