+++ src/main/java/kitchenpos/dto/request/table/TableIdRequest.java
@@ -1,0 +1,16 @@
+package kitchenpos.dto.request.table;
+
+public class TableIdRequest {
+    private Long id;
+
+    public TableIdRequest() {
+    }
+
+    public TableIdRequest(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+}
