+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ProductRepository;
-import kitchenpos.dto.request.CreateProductRequest;
+import kitchenpos.dto.request.product.CreateProductRequest;
 import kitchenpos.dto.response.ProductResponse;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
